requirements.txt
-i https://pypi.org/simple
cffi==1.14.0
cycler==0.10.0
fire==0.3.1
imageio==2.8.0
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
pillow==7.1.2
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.4.1
six==1.15.0
soundfile==0.10.3.post1
termcolor==1.1.0
